Got UniTask working with one acception: authentication
The Auth logic uses WhenAny and UniTask does not have a drop in
replacement. WhenAny works different for UniTask for some reason, and
I'm still looking for an answer.

Also, seems to be a bug on some playthroughs. Probably caused by
authentication issues or API calls to Cloud being dropped? Not sure.
Definitely buggy behavior, but at least a couple walkthroughs worked.